+++ AssetsFlow.Server/AssetsFlow.Journal/HsR.Journal.Entities/TradeJournal/TradeComposite.cs
@@ -2,31 +2,34 @@
 using HsR.Journal.Entities.TradeJournal;
 using System.ComponentModel.DataAnnotations;
 
-public class TradeComposite
+namespace HsR.Journal.Entities
 {
-    [Key]
-    public int Id { get; private set; }
-
-    [Required]
-    public Guid UserId { get; set; }
-
-    public ICollection<InterimTradeElement> TradeElements { get; set; } = new List<InterimTradeElement>();
-
-    public TradeSummary? Summary { get; set; }
-
-    #region Status Logic
-
-    public TradeStatus Status { get; private set; } = TradeStatus.AnIdea;
-
-    public void SetStatus(TradeStatus newStatus)
+    public class TradeComposite
     {
-        if (Status == newStatus)
-            return;
+        [Key]
+        public int Id { get; private set; }
 
-        Status = newStatus;
+        [Required]
+        public Guid UserId { get; set; }
+
+        public ICollection<InterimTradeElement> TradeElements { get; set; } = [];
+
+        public TradeSummary? Summary { get; set; }
+
+        #region Status Logic
+
+        public TradeStatus Status { get; private set; } = TradeStatus.AnIdea;
+
+        public void SetStatus(TradeStatus newStatus)
+        {
+            if (Status == newStatus)
+                return;
+
+            Status = newStatus;
+        }
+
+        #endregion
+
+        public override string ToString() => $"Id={Id}, Status={Status}, Trade Eles Count={TradeElements.Count}";
     }
-
-    #endregion
-
-    public override string ToString() => $"Id={Id}, Status={Status}, Trade Eles Count={TradeElements.Count}";
 }