+++ AssetsFlow.Server/AssetsFlow.Journal/HsR.Journal.Repository/Configurations/TradeCompositeConfig.cs
@@ -1,3 +1,4 @@
+using HsR.Journal.Entities;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;