+++ AssetsFlow.Server/AssetsFlow.Journal/HsR.Journal.Repository/Services/CompositeRepo/ITradeCompositeRepository.cs
@@ -1,4 +1,5 @@
-﻿using HsR.Journal.Services;
+﻿using HsR.Journal.Entities;
+using HsR.Journal.Services;
 using System;
 using System.Collections.Generic;
 using System.Linq;