+++ AssetsFlow.Server/AssetsFlow.Web.Services/Models/Journal/UpdatedStatesModel.cs
@@ -8,6 +8,7 @@ using System.Text;
 using System.Text.Json.Serialization;
 using System.Threading.Tasks;
 using System.ComponentModel.DataAnnotations;
+using AssetsFlowWeb.Services.Models.Journal;
 
 namespace HsR.Web.Services.Models.Journal
 {