+++ AssetsFlow.Server/Tests/HsR.Web.API.Caching.Tests/TradesCacheServiceTests.cs
@@ -5,7 +5,6 @@ using Microsoft.Extensions.Logging;
 using AutoMapper;
 using HsR.Web.API.Services;
 using HsR.Journal.DataContext;
-using HsR.Web.Services.Models.Journal;
 using HsR.Journal.Entities.TradeJournal;
 using System;
 using System.Collections.Generic;
@@ -13,6 +12,8 @@ using System.Threading.Tasks;
 using System.Threading;
 using Microsoft.Extensions.DependencyInjection;
 using HsR.Web.API.Settings;
+using AssetsFlowWeb.Services.Models.Journal;
+using HsR.Journal.Entities;
 
 namespace HsR.Web.API.Caching.Tests
 {