+++ AssetsFlow.Server/CHANGELOG.md
@@ -1,6 +1,8 @@
 
 ## [0.9.221-b] - 2025-07-...
 - Map date stamps on composite dto, removed double entries on composite models
+- DenyActivation flag added to content edit method
+- namespaces foxes
 
 ## [0.9.220-b] - 2025-07-26
 - User Service with gRPC + host + unittests